requirements.txt
streamlit
ultralytics
opencv-python
numpy
Pillow
torch
torchvision
--extra-index-url https://download.pytorch.org/whl/cpu
